[Security] Inherit COEP for newly created window
The spec states about COEP
"If creator is not null, set document’s embedder policy to creator
embedder policy.". This is done in this patch.

Note: This is required for COOP, because it cares about COEP in the
case of "same-origin" and needs it to be available on the initial empty
document of a new popup.

Also adds a WPT about a potential COEP exploit using WindowProxy to
circumvent COEP restrictions on fetching.

Bug:922191
Change-Id: Ie3097618be825c5031449d997970e33fad3d4f8e